chore: #94 move block getters to separate file
